@@ -12,37 +12,38 @@
   "author": "t-murota",
   "license": "MIT",
   "devDependencies": {
-    "babel": "^6.5.2",
-    "babel-loader": "^6.2.7",
+    "babel": "^6.23.0",
+    "babel-loader": "^7.1.5",
     "babel-preset-es2015": "^6.16.0",
     "babel-preset-react": "^6.16.0",
-    "css-loader": "^0.25.0",
-    "del": "^2.2.2",
+    "css-loader": "^1.0.0",
+    "del": "^3.0.0",
     "gulp": "^3.9.1",
-    "gulp-concat": "^2.6.0",
+    "gulp-concat": "^2.6.1",
     "gulp-env": "^0.4.0",
     "gulp-extend": "^0.2.0",
-    "gulp-header": "^1.8.8",
-    "gulp-imagemin": "^3.1.1",
+    "gulp-header": "^2.0.5",
+    "gulp-imagemin": "^4.1.0",
     "gulp-minify-html": "^1.0.6",
-    "gulp-plumber": "^1.1.0",
-    "gulp-replace": "^0.5.4",
-    "gulp-uglify": "^2.0.0",
-    "gulp-zip": "^3.2.0",
+    "gulp-plumber": "^1.2.0",
+    "gulp-replace": "^1.0.0",
+    "gulp-uglify": "^3.0.1",
+    "gulp-zip": "^4.2.0",
     "react-scripts": "0.4.1",
     "redux-devtools": "^3.3.1",
-    "run-sequence": "^1.2.2",
-    "style-loader": "^0.13.1",
-    "webpack": "^1.13.3",
-    "webpack-stream": "^3.2.0"
+    "run-sequence": "^2.2.1",
+    "style-loader": "^0.23.0",
+    "webpack": "^4.17.2",
+    "webpack-stream": "^5.1.1"
   },
   "dependencies": {
-    "moment": "^2.16.0",
+    "moment": "^2.22.2",
+    "prop-types": "^15.6.2",
     "react": "^15.3.2",
     "react-dom": "^15.3.2",
     "react-redux": "^4.4.5",
     "redux": "^3.6.0",
-    "redux-logger": "^2.7.4",
-    "redux-thunk": "^2.1.0"
+    "redux-logger": "^3.0.6",
+    "redux-thunk": "^2.3.0"
   }
 }
